requirements.txt
aniso8601==1.1.0
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
BeautifulSoup==3.2.1
blinker==1.4
certifi==2015.11.20.1
click==6.2
colorama==0.3.7
configparser==3.3.0.post2
cycler==0.10.0
dbutil==0.5.0
decorator==4.0.6
diff-match-patch==20121119
Django==1.8
django-daterange-filter==1.3.0
django-debug-toolbar==1.4
django-filter==0.14.0
django-filters==0.2.1
django-grappelli==2.7.2
django-grappelli-filters==0.2.1
django-import-export==0.4.5
django-jinja==2.1.1
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-session-security==2.2.4
django-simple-captcha==0.5.1
django-suit==0.2.15
django-utils==0.0.2
djangorestframework==3.4.6
djangorestframework-custom-filters==1.0.0
docopt==0.6.2
docutils==0.12
dominate==2.1.16
ecdsa==0.13
Flask==0.11.dev0
Flask-Bootstrap==3.3.5.7
Flask-Cache==0.13.1
Flask-CLI==0.2.1
Flask-DebugToolbar==0.10.0
Flask-HTTPAuth==2.7.0
Flask-Login==0.3.2
Flask-OpenID==1.2.5
flask-paginate==0.4.1
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
livereload==2.4.1
Markdown==2.6.5
markdown2==2.3.1
markup==0.2
MarkupSafe==0.23
matplotlib==1.5.1
MySQL-python==1.2.3
numpy==1.11.1
paramiko==1.16.0
passlib==1.6.5
pathlib2==2.1.0
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.7.4
Pillow==2.8.1
prompt-toolkit==1.0.7
ptyprocess==0.5.1
pycrypto==2.6
pydevd==0.0.5
Pygments==2.1
pyparsing==2.1.5
python-dateutil==2.4.2
python-openid==2.2.5
python-util==1.2.1
pytz==2015.7
pywin32==219
requests==2.9.1
simplegeneric==0.8.1
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.10.0
somewhere==0.0.1
SQLAlchemy==1.0.11
sqlalchemy-migrate==0.10.0
sqlparse==0.1.18
suit==2.0.2
tablib==0.11.2
Tempita==0.5.2
tornado==4.3
traitlets==4.2.2
visitor==0.1.2
wcwidth==0.1.7
web.py==0.38
WebOb==1.6.1
Werkzeug==0.11.3
win-unicode-console==0.5
WMI==1.4.9
WTForms==2.1
